Oops changed some of the wrong types

+++ day-8.rs
@@ -4,7 +4,7 @@ use std::collections::HashSet;
 fn main() {
     let file_str = fs::read_to_string("day-8.input").expect("Failed to read file");
 
-    let mut map : Vec<Vec<u32>> = Vec::new();
+    let mut map : Vec<Vec<u8>> = Vec::new();
 
     for line in file_str.trim().split('\n') {
         let mut row = Vec::new();
@@ -20,7 +20,7 @@ fn main() {
 
     map.iter().enumerate().fold(
         (0..map_width).map(|_| b'0' - 1).collect(),
-        |maxes : Vec<u32>, (x, row)| {
+        |maxes : Vec<u8>, (x, row)| {
             row.iter().zip(maxes).enumerate().map(|(y, (&height, prev_max))| {
                 if height > prev_max {
                     wins.insert((x as u32, y as u32));
@@ -33,7 +33,7 @@ fn main() {
 
     map.iter().enumerate().rev().fold(
         (0..map_width).map(|_| b'0' - 1).collect(),
-        |maxes : Vec<u32>, (x, row)| {
+        |maxes : Vec<u8>, (x, row)| {
             row.iter().zip(maxes).enumerate().map(|(y, (&height, prev_max))| {
                 if height > prev_max {
                     wins.insert((x as u32, y as u32));
